move transaction confirmation view to send folder
