+++ e2e/specs/confirmations/advanced-gas-fees.spec.js
@@ -3,7 +3,7 @@ import { SmokeConfirmations } from '../../tags';
 import WalletView from '../../pages/WalletView';
 import AmountView from '../../pages/Send/AmountView';
 import SendView from '../../pages/Send/SendView';
-import TransactionConfirmationView from '../../pages/TransactionConfirmView';
+import TransactionConfirmationView from '../../pages/Send/TransactionConfirmView';
 import { loginToApp } from '../../viewHelper';
 import FixtureBuilder from '../../fixtures/fixture-builder';
 import {
